fix: fix pixel sized visual artifacts
Previously, the pixels were sampled for texture at their top left
position. This caused weird visual artifacts, where a line
of pixels at the left edge was often left untextured. Also, it seems to
have caused some pixels to be left untextured along new triangles after
clipping.

To fix this, the texturing position of the pixels was changed to the
center of the pixel.