feat: add different camera styles
When panning the camera around with the mouse, mouse deltas must be
interpreted into camera rotations.

Previously, camera rotation was done locally, which meant that looking
right would pan the camera right aroudn the local y axis. This allowed
to camera to move freely independent of its rotation, but yielded
sometimes unwanted effects. Pressing shift before moving the mouse left
and right would also roll the camera.

FPS-like camera motion was added as an option when choosing camera
movement. This still pitches the camera with respect to the local x
axis, but yaw is done around the world aligned y axis. This results in
the local x axis of the camera always lying in the world's xz plane.
This essentially prevents any kind of roll.