feat: implement multithreaded logic
The method that will be used for running the rasterizer on multiple
cores has been implemented.

It has in no way been tested. This means there is absolutely no way that
thing is working at the moment. A lot of debugging will be necessary to
make this work. Even when it will, I am still unsure if it will increase
efficiency. The tiling overhead and memcpy at the end might overcome the
threading advantages. I guess we'll see!