fix: fix rasterizer's multithreading
The rasterizer now works using multiple cpu cores. It has not been
benchmarked with release mode yet, but my fps for a simple scene went
from ~15 to ~45 when close to the cube in debug mode. Release mode is
too fast and always caps my fps at 140. More tests will be required to
benchmark it.

Next, I which to introduce multiple objects to the scene to ensure
proper handling of overlapping shapes.