fix: allow multiple mesh to be rendered on screen
The single threaded rasterizer has been removed in favor of the
multithreaded one.

Also made the rasterizer into its own struct so that it can store the
pixel/depth tile buffers in memory. Before this, the buffers were
created in the rasterizer method and did not remember the previous
geometry. This prevented multiple meshes from appearing at once on the
scene.