feat: add shading
Flat shading for a light at infinity is now implemented. Done in an
extremely hacky way. At this point, the whole software renderer would be
in great need for a makeover to better structure the pipeline, make it
more understandle and more cache friendly. This feature exacerbated the
patchwork vibe of the project.

Remaining

Although available, Phong and Gouraud shading are not implemented.
Trying to use them will result in a crash. Given the extremely different
methods with which Flat, Gouraud and Phong shading work, I don't see an
efficient way of making all three available. Each require to be
introduced at a very specific point in the pipeline, which would either
require a specialied renderer for each shading technique, a shading map
(which would be expensive), or a match-if-else hell.

Same for the light point. It is available, but using it will result in a
crash.